Follow SOLID statements single responsibility. Changed some variables declaration to properties and it`s naming to PascalCasing
