+++ AppliancesModel/AppliancesStock.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AppliancesModel
+{
+    public enum AppliancesStock
+    {
+        Washer,
+        Refrigerator,
+        KitchenStove
+    }
+}
