+++ AppliancesModel/IOutputInputHandler.cs
@@ -1,0 +1,25 @@
+﻿namespace AppliancesModel
+{
+    public interface IOutputInputHandler
+    {
+        public void SelectApplianceToAdd(out int inputType, out int inputCount, string name);
+
+        public void SetApplianceProperties(out string name, out int guarantee, out Dimensions dimensions, out decimal price, out string producingCountry);
+
+        public void SetWasherProperties(out int waterConsuming, out int maximumLoad);
+
+        public void SetRefrigeratorProperties(out int totalVolume, out bool containsFreezer);
+
+        public void SetKitchenStoveProperties(out bool combinedGasElectric, out bool containsOven);
+
+        public void ShowRightsError();
+
+        public void ShowStockRefresh();
+
+        public void ShowStockNumbers(int washerCount, int refrigeratorCount, int kitchenStoveCount);
+
+        public int CheckIntInput(string article, int lowerBound, int upperBound);
+
+        public bool CheckBoolInput(string article);
+    }
+}
